Refactored duplicate functions into helper module
