+++ lib/qb_integration.rb
@@ -34,4 +34,3 @@ require 'qb_integration/services/invoice_line'
 require 'qb_integration/processors/invoice'
 require 'qb_integration/processors/invoice_line'
 require 'qb_integration/processors/customer'
-require 'qb_integration/processors/processor_helper'
