+++ lib/qb_integration/processors/customer.rb
@@ -1,6 +1,7 @@
 module QBIntegration
   module Processor
     class Customer
+      include Helper
       attr_reader :customer
 
       def initialize(customer)
@@ -52,31 +53,6 @@ module QBIntegration
 
       def email
         customer.primary_email_address ? customer.primary_email_address.address : ''
-      end
-
-      def build_address(addr)
-        return unless addr
-        {
-          id: addr["id"],
-          address1: addr["line1"],
-          address2: addr["line2"],
-          address3: addr["line3"],
-          address4: addr["line4"],
-          address5: addr["line5"],
-          city: addr["city"],
-          country: addr["country"],
-          state: addr["country_sub_division_code"],
-          country_sub_division_code: addr["country_sub_division_code"],
-          zipcode: addr["postal_code"],
-          note: addr["note"],
-          lat: addr["lat"],
-          lon: addr["lon"]
-        }
-      end
-
-      def build_ref(ref)
-        return {} unless ref
-        { name: ref["name"],id: ref["value"] }
       end
 
     end
