+++ lib/qb_integration/processors/invoice.rb
@@ -1,6 +1,7 @@
 module QBIntegration
   module Processor
     class Invoice
+      include Helper
       attr_reader :invoice
 
       def initialize(invoice)
@@ -74,12 +75,7 @@ module QBIntegration
       def build_email
         invoice.bill_email ? invoice.bill_email.address : ''
       end
-  
-      def build_ref(ref)
-        return {} unless ref
-        { name: ref["name"],id: ref["value"] }
-      end
-  
+      
       def build_tax_detail(tax)
         return {} unless tax
         {
@@ -120,26 +116,6 @@ module QBIntegration
             number_value: cf.number_value
           }
         end
-      end
-
-      def build_address(addr)
-        return {} unless addr
-        {
-          id: addr["id"],
-          address1: addr["line1"],
-          address2: addr["line2"],
-          address3: addr["line3"],
-          address4: addr["line4"],
-          address5: addr["line5"],
-          city: addr["city"],
-          country: addr["country"],
-          state: addr["country_sub_division_code"],
-          country_sub_division_code: addr["country_sub_division_code"],
-          zipcode: addr["postal_code"],
-          note: addr["note"],
-          lat: addr["lat"],
-          lon: addr["lon"]
-        }
       end
 
       # QuickBooks API doesn't provide status so we build the logic to determine status here
